+++ src/main/resources/archetype-resources/src/test/java/config/WebAppConfigurationAware.java
@@ -8,7 +8,11 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.context.WebApplicationContext;
 
 import javax.inject.Inject;
+import javax.xml.xpath.XPathExpressionException;
 
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.xpath;
 import static org.springframework.test.web.servlet.setup.MockMvcBuilders.webAppContextSetup;
 
 
@@ -29,4 +33,14 @@ public abstract class WebAppConfigurationAware {
         this.mockMvc = webAppContextSetup(this.wac).build();
     }
 
+    /**
+     * best way to check or validate xpath on Chrome browser is to use the Development tools JavaScript console and use the command 
+     * $x("/html/body//script[@src='/res/jquery/jquery.min.js']")
+     * @throws Exception to set
+     */
+    protected void testElementAvailable(String url, String xpath) throws XPathExpressionException, Exception {
+      mockMvc.perform(get(url).with(csrf().asHeader()))
+          // .andDo(print())
+          .andExpect(xpath(xpath).exists());
+    }
 }