+++ src/main/resources/archetype-resources/src/test/java/web/SigninControllerTest.java
@@ -4,7 +4,7 @@ import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.junit.jupiter.api.DisplayName;
@@ -12,6 +12,8 @@ import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.*;
+import static org.springframework.security.test.web.servlet.response.SecurityMockMvcResultMatchers.*;
+
 import ${package}.config.WebSecurityConfigurationAware;
 
 @Tag("controller")
@@ -24,7 +26,7 @@ public class SigninControllerTest extends WebSecurityConfigurationAware {
   @DisplayName("show signin form")
   public void displaySigninForm() throws Exception {
     mockMvc.perform(get("/signin").with(csrf().asHeader()))
-      .andDo(print())
+      .andDo(log()) // to activate change logger ".result" level to DEBUG in logback-test.xml
       .andExpect(status().isOk())
       .andExpect(view().name("signin/signin"))
       .andExpect(content().string(
@@ -43,9 +45,25 @@ public class SigninControllerTest extends WebSecurityConfigurationAware {
         .param("username", "user")
         .param("password", "user")
       )
-    .andDo(print())
+    .andDo(log()) // to activate change logger ".result" level to DEBUG in logback-test.xml
     .andExpect(status().isFound())
     .andExpect(redirectedUrl("/"))
+    .andExpect(authenticated().withUsername("user").withRoles("API", "USER"))
+    ;
+  }
+  
+  @Test
+  @DisplayName("proceed authenticate form wrong password")
+  public void signinFormWrongPassword() throws Exception {
+    mockMvc.perform(post("/authenticate")
+        .with(csrf().asHeader())
+        .param("username", "user")
+        .param("password", "invalid")
+      )
+    .andDo(log()) // to activate change logger ".result" level to DEBUG in logback-test.xml
+    .andExpect(status().isFound())
+    .andExpect(redirectedUrl("/signin?error=1"))
+    .andExpect(unauthenticated())
     ;
   }
 }