+++ changelog.md
@@ -7,7 +7,7 @@
 * remove maven profile test (useless with refactoring)
 * add configuration to deploy on Heroku for demo application
 * use @EventListener instead of implements ApplicationListener
-* increase test coverage to 45% with coveralls, implement HomeControllerTest, AboutControllerTest, SigninControllerTest
+* increase test coverage to 51% with coveralls, implement HomeControllerTest, AboutControllerTest, SigninControllerTest, ErrorControllerTest (total 21 tests)
 
 ## 5.1.5